Create Switchcase
Write a program that starts by displaying the following menu:


1. Find
2. Remove Duplicates
3. Merge strings
4. Reverse digits
5. Exit
- - - - - - - - - - - - -

Enter a choice:


Choice 1 will invoke a recursive method that accepts an array of integers and an integer i, and returns whether i is found in the array or not.


Choice 2 will invoke a recursive method that accepts a string s as a parameter and returns a string s’ with all consecutive duplicate characters in s removed (one remains). For example, if s=”aabbccded” then s’=”abcded”


Choice 3 will invoke a recursive method that accepts two strings s1 and s2 which contain words where the letters are alphabetically ordered (increasing order). The method returns a third string s3 with the letters in s1 and s2 merged in such a way that they are also sorted in s3. For example, if s1= “borx” and s2=”arst”, then s3=”aborrstx”.



Choice 4 will call the recursive method int reverseDigits(int n) which returns the positive integer that you get when you reverse the digits of parameter n. For example, when called with the parameter 12345, this method would return 54321. Do this with proper integer arithmetic instead of just simply converting to String, reversing that and then using parseInt () to extract the result.


Choice will terminate the program.


Any other choice, will have the menu displayed again.